Organizing the code
